+++ app/Http/Controllers/Message/MessageController.php
@@ -8,6 +8,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Message\MessageRequest;
 use App\Http\Resources\MessageResource;
 use App\Services\Message\MessageService;
+use Illuminate\Support\Facades\Gate;
 
 class MessageController extends Controller
 {
@@ -17,16 +18,22 @@ class MessageController extends Controller
 
     public function store(MessageRequest $request, Chat $chat)
     {
+        Gate::authorize('create', [Message::class, $chat]);
+
         return MessageResource::make($this->service->store($request, $chat));
     }
 
     public function update(MessageRequest $request, Chat $chat, Message $message)
     {
+        Gate::authorize('update', $message);
+
         return MessageResource::make($this->service->update($request, $message));
     }
 
     public function destroy(Message $message)
     {
+        Gate::authorize('delete', $message);
+
         $message->delete();
 
         return response()->noContent();
